# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NCST LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses/{courseId}/submission-slots:
    get:
      operationId: listSubmissionSlots
      tags: [submission-slots]
      summary: List submission slots for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionSlot"
    post:
      operationId: createSubmissionSlot
      tags: [submission-slots]
      summary: Create a submission slot (teacher/admin only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionSlotInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionSlot"
  /submission-slots/{id}:
    get:
      operationId: getSubmissionSlot
      tags: [submission-slots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionSlot"
    patch:
      operationId: updateSubmissionSlot
      tags: [submission-slots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionSlotInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionSlot"
    delete:
      operationId: deleteSubmissionSlot
      tags: [submission-slots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /submission-slots/{id}/submissions:
    get:
      operationId: listSlotSubmissions
      tags: [submission-slots]
      summary: List submissions for a slot (own only for students)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileSubmission"
...
components:
  schemas:
...
    FileSubmission:
      type: object
      required: [id, studentId, courseId, title, fileUrl, fileName, status, submittedAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        studentEmail:
          type: ["string", "null"]
        courseId:
          type: integer
        courseTitle:
          type: ["string", "null"]
        courseCode:
          type: ["string", "null"]
        slotId:
          type: ["integer", "null"]
        slotTitle:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        fileUrl:
          type: string
        fileName:
          type: string
        fileType:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, revision_requested]
        reviewerId:
          type: ["integer", "null"]
        reviewerName:
          type: ["string", "null"]
        reviewComment:
          type: ["string", "null"]
        submittedAt:
          type: string
          format: date-time
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SubmissionSlot:
      type: object
      required: [id, courseId, title, allowResubmission, createdBy, createdAt, updatedAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        dueAt:
          type: ["string", "null"]
          format: date-time
        allowResubmission:
          type: boolean
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        submissionCount:
          type: integer
        pendingCount:
          type: integer
        approvedCount:
          type: integer
        rejectedCount:
          type: integer
    SubmissionSlotInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        dueAt:
          type: ["string", "null"]
          format: date-time
        allowResubmission:
          type: boolean
...